refactor: change directory name
java code style 에 맞추어 디렉토리명을 소문자로만 구성하도록 변경
