+++ src/main/java/baseball/domain/hint/HintString.java
@@ -1,4 +1,4 @@
-package baseball.domain.Hint;
+package baseball.domain.hint;
 
 public enum HintString {
     BALL("%d볼 "),
